pyproject.toml
[project]
name = "devops-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "markdown (>=3.7,<4.0)",
    "django-filter (>=25.1,<26.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "django-safedelete (>=1.4.1,<2.0.0)",
    "django-simple-history (>=3.8.0,<4.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
